requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
readability-lxml==0.8.1
bleach==6.2.0
httpx==0.27.2
lxml==5.3.0
python-multipart==0.0.9
